+++ src/glyph_iterator.cpp
@@ -31,7 +31,7 @@ bool GlyphIterator::Valid() const
         return false;
     }
 
-    if (m_index < 0 || m_index >= m_pBuffer->GetWorkingBuffer().size())
+    if (m_index < 0 || static_cast<size_t>(m_index) >= m_pBuffer->GetWorkingBuffer().size())
     {
         return false;
     }
@@ -143,7 +143,7 @@ GlyphIterator& GlyphIterator::Move(long count)
     {
         for (long c = 0; c < count; c++)
         {
-            if (m_index < gapBuffer.size())
+            if (static_cast<size_t>(m_index) < gapBuffer.size())
             {
                 m_index += utf8_codepoint_length(gapBuffer[m_index]);
             }
